% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockwiseFusedLassoSignalApproximator.R
\name{blockwiseFusedLassoSignalApproximator}
\alias{blockwiseFusedLassoSignalApproximator}
\title{Blockwise fused lasso signal approximator}
\usage{
blockwiseFusedLassoSignalApproximator(indx, y, lambda)
}
\arguments{
\item{indx}{block indeces}

\item{y}{y}

\item{lambda}{\eqn{\lambda}}
}
\value{
\eqn{\beta}
}
\description{
Blockwise fused lasso signal approximator adapted from the glmgen library / package by
Taylor Arnold, Ryan Tibshirani, and Veerun Sadhanala.
}
\author{
Charles Kwong
}
