% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockwiseWeightedFusedLassoSignalApproximator.R
\name{blockwiseWeightedFusedLassoSignalApproximator}
\alias{blockwiseWeightedFusedLassoSignalApproximator}
\title{Blockwise weighted fused lasso signal approximator}
\usage{
blockwiseWeightedFusedLassoSignalApproximator(indx, y, w, lambda)
}
\arguments{
\item{indx}{block indeces}

\item{y}{y}

\item{w}{weights}

\item{lambda}{lambda}
}
\value{
beta
}
\description{
Blockwise weighted fused lasso signal approximator adapted from the glmgen library / package by
Taylor Arnold, Ryan Tibshirani, and Veerun Sadhanala.
}
\author{
Charles Kwong
}
