% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildFeatureMatrix.R
\name{buildFeatureMatrix}
\alias{buildFeatureMatrix}
\title{Feature matrix builder}
\usage{
buildFeatureMatrix(number_of_intervals, number_of_features,
  interval_numbers, feature_numbers, flags)
}
\arguments{
\item{number_of_intervals}{The number of intervals}
}
\value{
A matrix with `number_of_intervals` rows and `number_of_features` columns, where features are filled in with
1s according to the flags turning them on/off.
}
\description{
This builds the feature matrix out of flag vectors
(internally called function)
}
