% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdm_some_concept_names}
\alias{cdm_some_concept_names}
\title{A subset of the OMOP CDM "concept ancestor" table}
\format{a data frame (tibble)}
\source{
OMOP CDM Vocabulary tables
}
\usage{
data("omop_doi_concept_ancestor")
}
\description{
A subset of the OMOP CDM "concept ancestor" table
}
\keyword{datasets}
