% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkKKT4BetaStep.R
\name{checkKKT4BetaStep}
\alias{checkKKT4BetaStep}
\title{check convergence for beta step}
\usage{
checkKKT4BetaStep(y, X, w, beta, l, lambda)
}
\arguments{
\item{y}{response}

\item{X}{features}

\item{w}{weights}

\item{beta}{parameters learned}

\item{l}{interval length of each interval}

\item{lambda}{regularization strength}
}
\value{
2-norm of a vector that specifies the degree of
optimality violation
}
\description{
check conveergence for beta step
used as an inner loop convergence criterion
}
\author{
Zhaobin Kuang
}
