% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkKKT4BR.R
\name{checkKKT4BR}
\alias{checkKKT4BR}
\title{check convergence of baseline regularization (BR)}
\usage{
checkKKT4BR(Z, baseline_obs_period, X, l, n, t, beta, lambda1, lambda2,
  lambda3 = 0)
}
\arguments{
\item{Z}{design matrix for intervals}

\item{baseline_obs_period}{vector where index is the baseline parameter index and value is observation period index}

\item{X}{data matrix}

\item{l}{length of each interval}

\item{n}{number of dx occur at each interval}

\item{t}{baseline parameters}

\item{beta}{drug effect parameters}

\item{lambda1}{strength of ridge regularization for beta}

\item{lambda2}{strength of fused lasso regularization for baseline}

\item{lambda3}{strength of ridge regression for baseline}
}
\value{
2-norm of a vector that specifies the degree of
optimality violation
}
\description{
check the overall converge of BR
used as the outer loop stopping criterion
}
\author{
Zhaobin Kuang
}
