% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkKKT4MSCCS.R
\name{checkKKT4MSCCS}
\alias{checkKKT4MSCCS}
\title{check convergence of MSCCS}
\usage{
checkKKT4MSCCS(Z, X, y, w, lambda, alpha, beta)
}
\arguments{
\item{Z}{fixed effect matrix}

\item{X}{drug exposure matrix}

\item{y}{response count vector}

\item{w}{weight vector that describe data repetition}

\item{lambda}{regularization strength, single lambda}

\item{alpha}{fixed effect coefficients}

\item{beta}{covariate coefficents}
}
\value{
2-norm of a vector that specifies the degree of
optimality violation
}
\description{
check the convergence of MSCCS using KKT conditions
}
\author{
Zhaobin Kuang
}
