% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkL1Optimality.R
\name{checkL1Optimality}
\alias{checkL1Optimality}
\title{check whether l1 optimality}
\usage{
checkL1Optimality(subgradient, sgn)
}
\arguments{
\item{subgradient}{subgraident of l1-norm}

\item{sgn}{sign of the corresponding sub gradient}
}
\value{
2-norm whose value indicate optimality violation
}
\description{
check wehther a subgradient of a l1 penalty is optimal
}
