% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineBRParameters.R
\name{defineBRParameters}
\alias{defineBRParameters}
\title{Define Baseline Regularization parameters}
\usage{
defineBRParameters(lambda1 = 0, lambda2 = 0.5, lambda3 = 0.1,
  threshold = 1e-04, max_outer_loop_iterations = 10000L,
  max_inner_loop_iterations = 100000L, save_trajectory = F)
}
\arguments{
\item{lambda1}{Main effect L1 regularization strength}

\item{lambda2}{Fused lasso component regularization strength}

\item{lambda3}{Baseline parameters L2 regularization strength}

\item{threshold}{Threshold for breaking out of inner loop early}

\item{max_outer_loop_iterations}{Max outer loop iterations}

\item{max_inner_loop_iterations}{Max inner loop iterations}

\item{save_trajectory}{Whether to save all intermediate beta and t vectors (uses up lots of memory!)}
}
\value{
a parameters object representing these parameter settings
}
\description{
Define parameters for Baseline Regularization. See KDD 2017 paper for algorithm details.
}
