% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitBR.R
\name{fitBR}
\alias{fitBR}
\title{learn a baseline regularization (BR) model}
\usage{
fitBR(Z, interval_obs_period, X, l, n, lambda1, lambda2, lambda3 = 0,
  ...)
}
\arguments{
\item{Z}{design matrix that map tied baseline to
baseline on each interval}

\item{interval_obs_period}{a vector indicating to which observation period/patient each interval belongs.}

\item{X}{data matirx, each row represent data from
one interval, each column represent drug exposure info}

\item{l}{length of each interval}

\item{n}{number of condition occurrences on each interval}

\item{lambda1}{strength of ridge regression for beta}

\item{lambda2}{strengh of fused lasso for baseline}

\item{lambda3}{strength of ridge regression for baseline}

\item{...}{currently unused}
}
\value{
a br model
}
\description{
learn a BR model from matrix and vector input
currently only support ridge penalty onver drug effects
currently only support one pair of lambda1 and lambda2
learning at a time
}
\author{
Zhaobin Kuang
}
