% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMSCCS.R
\name{fitMSCCS}
\alias{fitMSCCS}
\title{Run MSCCS as an initialization for BR}
\usage{
fitMSCCS(obs_group, X, l, n, lambda, threshold = 1e-07)
}
\arguments{
\item{obs_group}{interval groups (observation periods)}

\item{X}{drug exposure matrix}

\item{l}{interval length}

\item{n}{number of response condition occurrences in an interval}

\item{lambda}{regularization strength, single lambda}

\item{threshold}{same as in glmnet}
}
\value{
an msccs model (list) that includes alpha, beta,
lambda, and err
}
\description{
Run MSCCS given matrix and vector input data from BR
currently only supports ridge regression over drug effects
currently only supports single lambda
this is used to initialze BR
}
\author{
Zhaobin Kuang
}
