% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAbsErr.R
\name{getAbsErr}
\alias{getAbsErr}
\title{Get absolute error between a pair of vectors/matrices}
\usage{
getAbsErr(a, b)
}
\arguments{
\item{a}{A vector/matrix}

\item{b}{A vector/matrix}
}
\value{
Frobenius norm between `a` and `b`.
}
\description{
Get absolute error between a pair of vectors/matrices.
Produces error messages if the objects have different dimensions.
}
\author{
Zhaobin Kuang

Yuriy Sverchkov
}
