% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoefficients.R
\name{getCoefficients}
\alias{getCoefficients}
\title{Get Beta Coefficients from fit}
\usage{
getCoefficients(fit, con = NULL, names_table = "concept",
  id_var = concept_id, name_var = concept_name, sort_by = desc(abs(Beta)),
  keep_zeros = F)
}
\arguments{
\item{fit}{the fit obtained from [fitBaselineRegularization].}

\item{con}{[DBI] connection for `names_table`.}

\item{names_table}{a table or the name of a database table (if con is provided) that matches the `concept_id` of the
drugs to more readable names.}

\item{id_var}{(NSE) the column in `names_table` that corresponds to drug `concept_id`.}

\item{name_var}{(NSE) the column in `names_table` that corresponds to drug name.}

\item{sort_by}{(NSE) the value by which to sort the beta coefficients.}

\item{keep_zeros}{whether to keep exact zeros in result.}
}
\value{
Table sorted as specified with columns Beta, concept_id, and (if possible) "Drug Name".
}
\description{
Function to conveniently get beta coefficients from the BR model in the form of a table, showing the concept_id and,
if available, name of the drug.
}
\author{
Yuriy Sverchkov
}
