% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConditionEvents.R
\name{getConditionEvents}
\alias{getConditionEvents}
\title{Get condition events from a condition occurrence/era table and an observation period table}
\usage{
getConditionEvents(condition, observation_period,
  date_column = !(!as.symbol("condition_start_date")))
}
\arguments{
\item{condition}{the condition occurrence or condition era table}

\item{observation_period}{the observation period table}

\item{date_column}{(NSE evaluation) the column in the condition table that should be used as the condition event time,
`condition_start_date` by default.}
}
\value{
a table of condition events
}
\description{
Get condition events from a condition occurrence/era table and an observation period table:
takes the start date to be the event time, merges on person_id with observation periods, and filters to keep the event
within the observation period.
}
\details{
Defaults work for the condition_occurrence table, need to specify date column for condition era
(`condition_era_start_date` according to OMOP CDM)
}
\author{
Yuriy Sverchkov
}
