% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDBTable.R
\name{getDBTable}
\alias{getDBTable}
\title{Get relevant tables from a database}
\usage{
getDBTable(con, table)
}
\arguments{
\item{con}{A DBI Connection obtained with DBI::dbConnect()}

\item{table}{The name of the table to extract}
}
\value{
The table
}
\description{
This is a weapper for dplyr::tbl that adds an explanatory flogger log message if dbplyr is not loaded.
}
