% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDrugDurationsFromExposure.R
\name{getDrugDurationsFromExposure}
\alias{getDrugDurationsFromExposure}
\title{Infer Drug Durations from the Drug Exposure Table}
\usage{
getDrugDurationsFromExposure(drug_exposure, observation_period)
}
\arguments{
\item{drug_exposure}{A dataframe-like object corresponding to the OMOP CDM `drug_exposure` table, particularly, we use
the columns `person_id`, `drug_concept_id`, `drug_exposure_start_date`, `drug_exposure_end_date`, and `days_supply`}

\item{observation_period}{A dataframe-like object corresponding to the OMOP CDM `observation_period` table}
}
\description{
Drug exposure durations are calculated as: exposure end date - exposure start date (if end date is specified) or
according to `days_supply` (if specified) or 1 day if neither is specified.
}
\author{
Yuriy Sverchkov
}
