% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDrugEvents.R
\name{getDrugEvents}
\alias{getDrugEvents}
\title{Get drug events from drug durations}
\usage{
getDrugEvents(drug_durations, risk_window = 0)
}
\arguments{
\item{drug_durations}{Drug durations table as returned by [getDrugDurationsFromExposure].}

\item{risk_window}{Risk window in days to be added to durations, with [Inf] indicating lasting risk.}
}
\description{
Takes a drug durations table, extends the durations by risk_window, merges overlapping durations, and retuens a
table of drug start and end events
}
\author{
Yuriy Sverchkov
}
