% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFusedSubgradient.R
\name{getFusedSubgradient}
\alias{getFusedSubgradient}
\title{get the subgradient of fused lasso}
\usage{
getFusedSubgradient(grad, lambda)
}
\arguments{
\item{grad}{gradient of the original fused lasso regularized problem}

\item{lambda}{strength of fused lasso regularization, strictly positive}
}
\value{
fused lasso subgradient
}
\description{
get the subgraident of fused lasso given gradient and lambda
}
\author{
Zhaobin Kuang
}
