% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWls.R
\name{getWls}
\alias{getWls}
\title{compute weighted least square}
\usage{
getWls(y, X, w, lambda, thresh = 1e-07)
}
\arguments{
\item{y}{response}

\item{X}{features}

\item{w}{weights}

\item{lambda}{ridge regularization strength}

\item{thresh}{defaut to be 1e-7, convergence criterion of
glmnet}
}
\value{
beta the regression coefficient
}
\description{
solve a weighted least square problem
it can be reugularized by a ridge penaly
currently only support one lambda at a time
}
\author{
Zhaobin Kuang
}
