% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferObservationPeriods.R
\name{inferObservationPeriods}
\alias{inferObservationPeriods}
\title{Construct an observation period table from a set of tables}
\usage{
inferObservationPeriods(..., patient_id = patient_id,
  time_match = "_date$")
}
\arguments{
\item{...}{One or more tables}

\item{patient_id}{specifies the column that corresponds to the patient ID (each observation period has only one of
these), using non-standard evaluation}

\item{time_match}{regular expression to match to columns that correspond to event times}
}
\value{
a table with three columns: patient_id, observation_period_start, observation_period_end
}
\description{
Takes one or more tables and returns a table of 'patient' observation periods basd on the first and last dates
that are associated with the patients
}
