% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isSingleString.R
\name{isSingleString}
\alias{isSingleString}
\title{Check if an object is a single string}
\usage{
isSingleString(x)
}
\arguments{
\item{x}{the object to check.}
}
\value{
TRUE iff x is of type character and of length 1.
}
\description{
Check if an object is a single string
}
\author{
Yuriy Sverchkov
}
