% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{omop_concept_names}
\alias{omop_concept_names}
\title{A subset of the "concept" table for OMOP concepts}
\format{a data frame (tibble)}
\source{
OMOP CDM Vocabulary tables
}
\usage{
data("omop_doi_concept_ancestor")
}
\description{
A subset of the "concept" table for OMOP concepts
}
\keyword{datasets}
