% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{omop_doi_concept_ancestor}
\alias{omop_doi_concept_ancestor}
\title{A subset of the "concept ancestor" table for OMOP concepts}
\format{a data frame (tibble) with columns `descendant_concept_id` and `ancestor_concept_id`}
\source{
OMOP CDM Vocabulary tables
}
\usage{
data("omop_doi_concept_ancestor")
}
\description{
A subset of the "concept ancestor" table for OMOP concepts
}
\keyword{datasets}
