% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{omop_hoi_conditions_map}
\alias{omop_hoi_conditions_map}
\title{A mapping from OMOP HOIs to condition concept IDs}
\format{a data frame (tibble)}
\source{
OMOP CDM Vocabulary tables and HOI definitions
}
\usage{
data("omop_doi_concept_ancestor")
}
\description{
A mapping from OMOP HOIs to condition concept IDs
}
\keyword{datasets}
