% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passthroughConceptProcessor.R
\name{passthroughConceptProcessor}
\alias{passthroughConceptProcessor}
\title{Passthrough concept processor}
\usage{
passthroughConceptProcessor(record_table,
  record_table_column = "concept_id", out_column = record_table_column)
}
\arguments{
\item{record_table}{The table to process}

\item{record_table_column}{The record table column to process}

\item{out_column}{The column to which to write the processor result}
}
\value{
`record_table` with `out_column` added, being a copy of `record_table_column`
}
\description{
A concept processor that just copies `record_table_column` to `out_column`
}
