% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareBRData.R
\name{prepareBRData}
\alias{prepareBRData}
\title{Create data object based on era tables}
\usage{
prepareBRData(con = NULL, observation_period = "observation_period",
  drug_era = "drug_era", condition_era = "condition_era",
  drug_exposure = "drug_exposure",
  condition_occurrence = "condition_occurrence",
  visit_occurrence = "visit_occurrence", response_event,
  tying = "occurence", risk_window = 0L, minimum_duration = 0L,
  drug_concept_processor = passthroughConceptProcessor,
  condition_concept_processor = passthroughConceptProcessor)
}
\arguments{
\item{con}{A DBI Connection to an OMOP CDM database used if the OMOP tables are provided as table names.}

\item{observation_period}{Either the name of the `observation_period` table in the database specified by
`con` or the table itself in a dataframe-like R object. If this is specified it defines the observation periods; if
it is not, the observation periods are inferred from the other tables in the data, assuming one all-encompassing
observation period per patient.}

\item{drug_era}{A dataframe-like view or the table name of the OMOP DRUG_ERA table}

\item{condition_era}{A dataframe-like view or the table name of the OMOP CONDITION_ERA table}

\item{drug_exposure}{Either the name of the `drug_exposure` table in the database specified by `con` or the
table itself in a dataframe-like R object}

\item{condition_occurrence}{Either the name of the `condition_occurrence` table in the database specified by `con`
or the table itself in a dataframe-like R object}

\item{visit_occurrence}{(optional) Either the name of the `visit_occurrence` table in the database specified by `con`
or the table itself in a dataframe-like R object, only used to infer observation periods if needed.}

\item{response_event}{The condition_concept_id of the condition of interest}

\item{tying}{Parameter tying mode, "interval", or "occurence" (default)}

\item{risk_window}{The number of days right after a drug era during which the patient is considered still under
exposure.}

\item{minimum_duration}{The number of days a patient must be under observation to be included in the analysis.}
}
\value{
An object containing the matrices X, Z, y
}
\description{
Creates the data object for running Baseline Regularization
}
\author{
Yuriy Sverchkov
}
