% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareBRDataFromEras.R
\name{prepareBRDataFromEras}
\alias{prepareBRDataFromEras}
\title{Create data object}
\usage{
prepareBRDataFromEras(con = NULL, observation_period = NULL,
  drug_era = NULL, condition_era = NULL, event, tying = "occurence",
  risk_window = 0, minimum_duration = 0,
  independent_observation_periods = TRUE)
}
\arguments{
\item{con}{A DBI Connection to an OMOP CDM database. If it is provided, the observation_period, drug_era, and
condition_era tables will be read from the database and the corresponding parameters provided to this function will
be ignored. If it is not provided, the tables will be read from the parameters.}

\item{observation_period}{A dataframe-like view of the OMOP OBSERVATION_PERIOD table}

\item{drug_era}{A dataframe-like view of the OMOP DRUG_ERA table}

\item{condition_era}{A dataframe-like view of the OMOP CONDITION_ERA table}

\item{tying}{Parameter tying mode, "interval", or "occurence" (default)}

\item{risk_window}{The number of days right after a drug era during which the patient is considered still under
exposure.}

\item{minimum_duration}{The number of days a patient must be under observation to be included in the analysis.}

\item{independent_observation_periods}{Whether to treat distinct observation periods from one patient as distinct
"patients." Default value and only current accepted value is TRUE.}

\item{condition}{The condition_concept_id of the condition of interest}
}
\value{
An object containing the matrices X, Z, y
# maybe import tidyr
}
\description{
Creates the data object for running Baseline Regularization
}
