% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareBRDataFromEvents2.R
\name{prepareBRDataFromEvents2}
\alias{prepareBRDataFromEvents2}
\title{Generate BR data from event table, (SQLite-friendly internal function)}
\usage{
prepareBRDataFromEvents2(all_events, event, tying)
}
\arguments{
\item{all_events}{Table of events with columns:
`obs_period_id`,
`event_day`,
`concept_id`,
`event_flag`,
`observation_period_length`.}

\item{event}{The event of interest the risk of which to estimate.}

\item{tying}{The type of tying to use (`interval` or `occurrence`).}
}
\value{
Data for [fitBaselineRegularization].
}
\description{
Internal function that prepares data for [fitBaselineRegularization] from the output of a `getEventsFrom*` function.
This version specitically written to work with SQLite.
}
\author{
Yuriy Sverchkov
}
