% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareBRDataFromOccurrence.R
\name{prepareBRDataFromOccurrence}
\alias{prepareBRDataFromOccurrence}
\title{Create data object for baseline regularization}
\usage{
prepareBRDataFromOccurrence(con = NULL, drug_exposure = "drug_exposure",
  condition_occurrence = "condition_occurrence",
  visit_occurrence = "visit_occurrence", event, tying = "occurrence",
  risk_window = 0, minimum_duration = 0)
}
\arguments{
\item{con}{A DBI Connection to an OMOP CDM database. If this is absent or NULL, tables must be provided as
dataframe-like R objects.}

\item{drug_exposure}{Either the name of the `drug_exposure` table in the database specified by `con` or the
table itself in a dataframe-like R object}

\item{condition_occurrence}{Either the name of the `condition_occurrence` table in the database specified by `con`
or the table itself in a dataframe-like R object}

\item{visit_occurrence}{Either the name of the `visit_occurrence` table in the database specified by `con` or the
table itself in a dataframe-like R object.
A patient will be considered observed from their first to their last visit.
If this parameter is NULL then the patient will be considered from their first to their last drug exposure and/or
condition occurence}

\item{event}{The condition_concept_id of the event of interest}

\item{tying}{Parameter tying mode, "interval", or "occurence" (default)}

\item{risk_window}{The number of days right after a drug exposure during which the patient is considered still under
exposure.}

\item{minimum_duration}{The number of days a patient must be under observation to be included in the analysis.}
}
\value{
An object containing the matrices X, Z, y
}
\description{
Creates the data object for running Baseline Regularization
}
