% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synpuf_mini}
\alias{synpuf_mini}
\title{A mini subset of the synpuf dataset in OMOP CDM}
\format{a list of dataframes: `observation_period`, `condition_occurrence`, `drug_exposure`}
\source{
The CMS 2008-2010 Data Entrepreneurs’ Synthetic Public Use File (DE-SynPUF)
https://www.cms.gov/Research-Statistics-Data-and-Systems/Downloadable-Public-Use-Files/SynPUFs/DE_Syn_PUF.html
}
\usage{
data("synpuf_mini")
}
\description{
A mini subset of the CMS 2008-2010 Data Entrepreneurs’ Synthetic Public Use File (DE-SynPUF) dataset
in OMOP CDM V.5.
}
\keyword{datasets}
