% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineBRParameters.R
\name{defineBRParameters}
\alias{defineBRParameters}
\title{Define Baseline Regularization parameters}
\usage{
defineBRParameters(lambda1 = 0, lambda2 = 0.5, lambda3 = 0.1,
  threshold = 1e-04, maxOuterLoopIterations = 60,
  maxInnerLoopIterations = 200)
}
\arguments{
\item{lambda1}{Main effect L1 regularization strength}

\item{lambda2}{Fused lasso component regularization strength}

\item{lambda3}{Baseline parameters L2 regularization strength}

\item{threshold}{Threshold for breaking out of inner loop early}

\item{maxOuterLoopIterations}{Max outer loop iterations}

\item{maxInnerLoopIterations}{Max inner loop iterations}
}
\value{
a parameters object representing these parameter settings
}
\description{
Define parameters for Baseline Regularization. See KDD 2017 paper for algorithm details.
}
