% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitBaselineRegularization.R
\name{fitBaselineRegularization}
\alias{fitBaselineRegularization}
\title{Fit baseline regularization}
\usage{
fitBaselineRegularization(brData, parameters = defineBRParameters())
}
\arguments{
\item{brData}{data object produced by prepareBRData}

\item{parameters}{parameters object produced by defineBRParameters}
}
\value{
An object representing the fit model, consisting of the main effect weights and baseline biases
}
\description{
Fit Baseline Regularization with the given parameters
}
