% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDBTable.R
\name{getDBTable}
\alias{getDBTable}
\title{Get relevant tables from a database}
\usage{
getDBTable(con, table, label = NULL)
}
\arguments{
\item{con}{A DBI Connection obtained with DBI::dbConnect()}

\item{table}{The name of the table to extract}
}
\value{
The table
}
\description{
Takes a connection, table or table name, and (optional) label, if the second parameter isn't a table name, it just
returns it, otherwise, announces that it is loading the table from the database and uses [dplyr::tbl] to do that.
Also adds an explanatory message if dbplyr is not loaded.
}
