% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEventsFromOccurrence.R
\name{getEventsFromOccurrence}
\alias{getEventsFromOccurrence}
\title{Make events table from occurrence information (internal function)}
\usage{
getEventsFromOccurrence(drug_exposure, condition_occurrence,
  observation_periods, event, risk_window, drug_sym, condition_sym,
  person_sym)
}
\arguments{
\item{drug_exposure}{A dataframe-like object corresponding to the OMOP CDM `drug_exposure` table}

\item{condition_occurrence}{A dataframe-like object corresponding to the OMOP CDM `condition_occurrence` table}

\item{observation_periods}{A dataframe-like object corresponding to the OMOP CDM `observation_period` table}

\item{event}{The event of interest}

\item{risk_window}{See note on drug exposure duration calculation}

\item{drug_sym}{A [name] representing the column to use for drug ids, drug_concept_id by default.}

\item{condition_sym}{A [name] representing the column to use for condition ids, condition_concept_id by default.}

\item{person_sym}{A [name] representing the column to use for person ids, person_id by default.}
}
\description{
Make events table for [prepareBRDataFromEvents] from occurrence tables (rather than era tables).
}
\details{
Drug exposure durations are calculated as: exposure end date - exposure start date (if end date is specified) or
according to `days_supply` (if specified) or 1 day if neither is specified, plus `risk_window`.
}
\author{
Yuriy Sverchkov
}
