% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareBRDataFromEras.R
\name{prepareBRDataFromEras}
\alias{prepareBRDataFromEras}
\title{Create data object based on era tables}
\usage{
prepareBRDataFromEras(con = NULL, observation_period = "observation_period",
  drug_era = "drug_era", condition_era = "condition_era", response_event,
  tying = "occurence", risk_window = 0, minimum_duration = 0,
  independent_observation_periods = TRUE)
}
\arguments{
\item{con}{A DBI Connection to an OMOP CDM database used if the OMOP tables are provided as table names.}

\item{observation_period}{A dataframe-like view or the table name of the OMOP OBSERVATION_PERIOD table}

\item{drug_era}{A dataframe-like view or the table name of the OMOP DRUG_ERA table}

\item{condition_era}{A dataframe-like view or the table name of the OMOP CONDITION_ERA table}

\item{response_event}{The condition_concept_id of the condition of interest}

\item{tying}{Parameter tying mode, "interval", or "occurence" (default)}

\item{risk_window}{The number of days right after a drug era during which the patient is considered still under
exposure.}

\item{minimum_duration}{The number of days a patient must be under observation to be included in the analysis.}

\item{independent_observation_periods}{Whether to treat distinct observation periods from one patient as distinct
"patients." Default value and only current accepted value is TRUE.}
}
\value{
An object containing the matrices X, Z, y
}
\description{
Creates the data object for running Baseline Regularization
}
\author{
Yuriy Sverchkov
}
