% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synpuf_mini}
\alias{synpuf_mini}
\title{A mini subset of the synpuf dataset in OMOP CDM}
\format{a list of dataframes: `observation_period`, `condition_occurrence`, `drug_exposure`}
\source{
synpuf (TODO: fill out)
}
\usage{
data(synpud_mini)
}
\description{
A mini subset of the synpuf dataset in OMOP CDM
}
\references{
TODO: fill out
}
\keyword{datasets}
