% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useAncestorConcepts.R
\name{useAncestorConcepts}
\alias{useAncestorConcepts}
\title{Use ancestor concepts}
\usage{
useAncestorConcepts(record_table, ancestors,
  record_table_column = "concept_id",
  ancestor_column = "ancestor_concept_id",
  descendant_column = "descendant_concept_id", copy = FALSE)
}
\arguments{
\item{record_table}{The table to augment}

\item{ancestors}{The table defining ancestry relations}

\item{record_table_column}{The column in `record_table` to be matched to ancestors}

\item{ancestor_column}{The column with which to augment `record_table`}

\item{descendant_column}{The column which to match to `record_table_column`}

\item{copy}{Passthrough parameter to [dplyr::left_join], needed when `record_table` and `ancestors` have different
sources}
}
\description{
Use an ancestry table to add a column of ancestor concepts to a record table
This is basically a wrapper for [dplyr::left_join], but only keeps one column of interest from the ancestry table
}
\author{
Yuriy Sverchkov
}
